+++ src/uball.h
@@ -4,33 +4,33 @@
 class Cball {
     typedef short gene[140];
   public:
-    Cball (short x, short y, short s)
+    Cball (int x, int y, int s)
     : xp(x), yp(y), ix(x), iy(y), sexi(s),
       nomating(0), nocandiate(0),
       fitness(0), dfitness(0), mdistance(0) {}
     // initialize genes
-    void Igene(short nthgene, short gf, short gl) {
+    void Igene(int nthgene, short gf, short gl) {
          gene1[nthgene] = gf;
          gene2[nthgene] = gl;
          genotype[nthgene] = gf + gl;
     }
     double ResourceM() const;
-    void nreproduction(const Cball& male, std::list<Cball>* ablist, short nogene, double mdis, double fdis, double mr, double nmr);
+    void nreproduction(const Cball& male, std::list<Cball>* ablist, int nogene, double mdis, double fdis, double mr, double nmr);
     void measurefitness(double RR, double gradient, double Vs, double K, double Range, double MS);
 
-    short xp, yp, ix, iy, sexi;
-    short nomating, nocandiate;
+    int xp, yp, ix, iy, sexi;
+    int nomating, nocandiate;
     double fitness, dfitness, mdistance;
     gene gene1, gene2, genotype;
     std::list<Cball>::iterator candidatemate[1500];
 };
 
 void Newball(std::list<Cball>* list1);
-void Newball2008(short n, short male, std::list<Cball>* list1, double fr);
-short matingcount(std::list<Cball>::iterator focalindiv, std::list<Cball>::iterator* matp, short matingsize);
-void SaveF(const std::list<Cball>& clist, short g, short gg, long n, short nogene);
+void Newball2008(int n, int male, std::list<Cball>* list1, double fr);
+int matingcount(std::list<Cball>::iterator focalindiv, std::list<Cball>::iterator* matp, int matingsize);
+void SaveF(const std::list<Cball>& clist, int g, int gg, size_t n, int nogene);
 short Resource(short a, short b);
-void SaveE(short g, short gg);
-void SaveA(const std::list<Cball>& clist, short g, short gg, long n, short clas );
+void SaveE(int g, int gg);
+void SaveA(const std::list<Cball>& clist, int g, int gg, size_t n, int clas);
 void AssignBucket(std::list<Cball>* list1);
 //Individuals is stored in each bucket according to the position.
